pyproject.toml
[project]
name = "documind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncpg>=0.31.0",
    "chromadb>=1.5.2",
    "fastapi>=0.133.0",
    "google>=3.0.0",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
    "huggingface-hub>=1.5.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "loguru>=0.7.3",
    "mistralai>=1.12.4",
    "opencv-python>=4.13.0.92",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.3",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
